Move deployment code out of admin client package
