+++ java/clients/src/main/java/sleeper/clients/deploy/CdkDeployInstance.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.clients.admin.deploy;
+package sleeper.clients.deploy;
 
 import sleeper.configuration.properties.InstanceProperties;
 import sleeper.util.ClientUtils;
