+++ java/clients/src/test/java/sleeper/clients/deploy/GeneratePropertiesTest.java
@@ -13,27 +13,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.clients.admin.deploy;
 
-import com.amazonaws.services.securitytoken.AWSSecurityTokenService;
-import com.amazonaws.services.securitytoken.AWSSecurityTokenServiceClientBuilder;
-import com.amazonaws.services.securitytoken.model.GetCallerIdentityRequest;
+package sleeper.clients.deploy;
+
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.api.io.TempDir;
-import org.testcontainers.containers.localstack.LocalStackContainer;
-import org.testcontainers.junit.jupiter.Container;
-import org.testcontainers.junit.jupiter.Testcontainers;
-import org.testcontainers.utility.DockerImageName;
 import software.amazon.awssdk.regions.Region;
 
 import sleeper.configuration.properties.InstanceProperties;
-import sleeper.configuration.properties.local.SaveLocalProperties;
 import sleeper.configuration.properties.table.TableProperties;
-import sleeper.core.CommonTestConstants;
-
-import java.io.IOException;
-import java.nio.file.Path;
-import java.util.stream.Stream;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static sleeper.configuration.properties.SystemDefinedInstanceProperty.CONFIG_BUCKET;
@@ -47,27 +34,16 @@ import static sleeper.configuration.properties.UserDefinedInstanceProperty.JARS_
 import static sleeper.configuration.properties.UserDefinedInstanceProperty.REGION;
 import static sleeper.configuration.properties.UserDefinedInstanceProperty.SUBNET;
 import static sleeper.configuration.properties.UserDefinedInstanceProperty.VPC_ID;
+import static sleeper.configuration.properties.table.TableProperty.DATA_BUCKET;
+import static sleeper.configuration.properties.table.TableProperty.TABLE_NAME;
 import static sleeper.core.schema.SchemaTestHelper.schemaWithKey;
 
-@Testcontainers
-public class GeneratePropertiesIT {
-
-    @Container
-    public static LocalStackContainer localStackContainer = new LocalStackContainer(DockerImageName.parse(CommonTestConstants.LOCALSTACK_DOCKER_IMAGE))
-            .withServices(LocalStackContainer.Service.S3, LocalStackContainer.Service.STS);
-    private final AWSSecurityTokenService sts = AWSSecurityTokenServiceClientBuilder.standard()
-            .withEndpointConfiguration(localStackContainer.getEndpointConfiguration(LocalStackContainer.Service.STS))
-            .withCredentials(localStackContainer.getDefaultCredentialsProvider())
-            .build();
-
-    @TempDir
-    private Path tempDir;
-
+public class GeneratePropertiesTest {
     @Test
     void shouldGenerateInstancePropertiesCorrectly() {
         // Given/When
         InstanceProperties properties = generateInstancePropertiesBuilder()
-                .sts(sts).regionProvider(() -> Region.of(localStackContainer.getRegion()))
+                .instanceId("test-instance").vpcId("some-vpc").subnetId("some-subnet")
                 .build().generate();
 
         // Then
@@ -81,44 +57,44 @@ public class GeneratePropertiesIT {
         expected.set(ECR_COMPACTION_REPO, "test-instance/compaction-job-execution");
         expected.set(ECR_INGEST_REPO, "test-instance/ingest");
         expected.set(BULK_IMPORT_REPO, "test-instance/bulk-import-runner");
-        expected.set(ACCOUNT, sts.getCallerIdentity(new GetCallerIdentityRequest()).getAccount());
-        expected.set(REGION, localStackContainer.getRegion());
+        expected.set(ACCOUNT, "test-account-id");
+        expected.set(REGION, "aws-global");
 
         assertThat(properties).isEqualTo(expected);
     }
 
     @Test
-    void shouldSaveBucketNamesToLocalDirectoryWhenInstancePropertiesGenerated() throws IOException {
-        // Given
-        InstanceProperties properties = generateInstancePropertiesBuilder()
-                .sts(sts).regionProvider(() -> Region.of(localStackContainer.getRegion()))
-                .build().generate();
-
-        // When
-        SaveLocalProperties.saveToDirectory(tempDir, properties, Stream.empty());
-
-        // Then
-        assertThat(tempDir.resolve("configBucket.txt")).exists();
-        assertThat(tempDir.resolve("queryResultsBucket.txt")).exists();
-    }
-
-    @Test
-    void shouldSaveBucketNamesToLocalDirectoryWhenTablePropertiesGenerated() throws IOException {
+    void shouldGenerateTablePropertiesCorrectly() {
         // Given
         InstanceProperties instanceProperties = generateInstancePropertiesBuilder()
-                .sts(sts).regionProvider(() -> Region.of(localStackContainer.getRegion()))
+                .instanceId("test-instance").vpcId("some-vpc").subnetId("some-subnet")
                 .build().generate();
         TableProperties tableProperties = GenerateTableProperties.from(instanceProperties, schemaWithKey("key"), "test-table");
 
-        // When
-        SaveLocalProperties.saveToDirectory(tempDir, instanceProperties, Stream.of(tableProperties));
-
         // Then
-        assertThat(tempDir.resolve("tables/test-table/tableBucket.txt")).exists();
+        InstanceProperties expectedInstanceProperties = new InstanceProperties();
+        expectedInstanceProperties.set(ID, "test-instance");
+        expectedInstanceProperties.set(CONFIG_BUCKET, "sleeper-test-instance-config");
+        expectedInstanceProperties.set(JARS_BUCKET, "sleeper-test-instance-jars");
+        expectedInstanceProperties.set(QUERY_RESULTS_BUCKET, "sleeper-test-instance-query-results");
+        expectedInstanceProperties.set(VPC_ID, "some-vpc");
+        expectedInstanceProperties.set(SUBNET, "some-subnet");
+        expectedInstanceProperties.set(ECR_COMPACTION_REPO, "test-instance/compaction-job-execution");
+        expectedInstanceProperties.set(ECR_INGEST_REPO, "test-instance/ingest");
+        expectedInstanceProperties.set(BULK_IMPORT_REPO, "test-instance/bulk-import-runner");
+        expectedInstanceProperties.set(ACCOUNT, "test-account-id");
+        expectedInstanceProperties.set(REGION, "aws-global");
+
+        TableProperties expected = new TableProperties(expectedInstanceProperties);
+        expected.setSchema(schemaWithKey("key"));
+        expected.set(TABLE_NAME, "test-table");
+        expected.set(DATA_BUCKET, "sleeper-test-instance-table-test-table");
+
+        assertThat(tableProperties).isEqualTo(expected);
     }
 
     private GenerateInstanceProperties.Builder generateInstancePropertiesBuilder() {
         return GenerateInstanceProperties.builder()
-                .instanceId("test-instance").vpcId("some-vpc").subnetId("some-subnet");
+                .accountSupplier(() -> "test-account-id").regionProvider(() -> Region.AWS_GLOBAL);
     }
 }
