+++ java/clients/src/test/java/sleeper/clients/deploy/UploadDockerImagesTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package sleeper.clients.admin.deploy;
+package sleeper.clients.deploy;
 
 import org.junit.jupiter.api.Test;
